chore: move info to db info
