+++ chroma_ops/main.py
@@ -2,7 +2,7 @@ import typer
 
 from chroma_ops.fts import fts_commands
 from chroma_ops.wal import wal_commands
-from chroma_ops.info import command as info_command
+from chroma_ops.db import db_commands
 from chroma_ops.clean import command as clean_command
 from chroma_ops.hnsw import hnsw_commands
 from chroma_ops.collection import collection_commands
@@ -12,18 +12,14 @@ app = typer.Typer(no_args_is_help=True, help="ChromaDB Ops Commands.")
 
 app.add_typer(wal_commands, name="wal", help="WAL maintenance commands")
 
-app.command(
-    name="info",
-    help="Provide persistent Chroma DB information. "
-    "Useful to understand how your Chroma works or get support from the team.",
-    no_args_is_help=True,
-)(info_command)
 
 app.command(
     name="clean",
     help="Clean up orphaned vector segment directories.",
     no_args_is_help=True,
 )(clean_command)
+
+app.add_typer(db_commands, name="db", help="DB maintenance commands")
 
 app.add_typer(hnsw_commands, name="hnsw", help="HNSW index maintenance commands")
 app.add_typer(
